# Use recycled buffer when serializaing trace header.