--- a/aws-xray-recorder-sdk-benchmark/tst/main/java/com/amazonaws/xray/entities/TraceHeaderBenchmark.java
+++ b/aws-xray-recorder-sdk-benchmark/tst/main/java/com/amazonaws/xray/entities/TraceHeaderBenchmark.java
@@ -35,18 +35,24 @@ import org.openjdk.jmh.runner.options.OptionsBuilder;
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 public class TraceHeaderBenchmark {
 
-    private static final String HEADER = "Root=1-8a3c60f7-d188f8fa79d48a391a778fa6;Parent=53995c3f42cd8ad8;Sampled=1";
+    private static final String HEADER_STRING = "Root=1-8a3c60f7-d188f8fa79d48a391a778fa6;Parent=53995c3f42cd8ad8;Sampled=1";
+    private static final TraceHeader HEADER = TraceHeader.fromString(HEADER_STRING);
 
     @Benchmark
     public TraceHeader parse() {
-        return TraceHeader.fromString(HEADER);
+        return TraceHeader.fromString(HEADER_STRING);
+    }
+
+    @Benchmark
+    public String serialize() {
+        return HEADER.toString();
     }
 
     // Convenience main entry-point
     public static void main(String[] args) throws RunnerException {
         Options opt = new OptionsBuilder()
             .addProfiler("gc")
-            .include(".*" + TraceHeaderBenchmark.class.getSimpleName())
+            .include(".*" + TraceHeaderBenchmark.class.getSimpleName() + ".serialize")
             .build();
 
         new Runner(opt).run();