--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/internal/RecyclableBuffers.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/internal/RecyclableBuffers.java
@@ -40,5 +40,6 @@ public final class RecyclableBuffers {
         return buffer;
     }
 
-    private RecyclableBuffers() {}
+    private RecyclableBuffers() {
+    }
 }