Many libraries, like Jackson, logging, and other tracing libraries, use a `ThreadLocal` buffer when creating strings. This is because Java will always copy the buffer when creating the `String` so the work of allocating / zeroing an intermediate buffer for filling up the content is completely wasted. 

Using recyclable buffers requires a bit of thought but don't add real complexity to the code so I think they're generally worth it.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

About 20% speedup
After
```
Benchmark                                                          Mode     Cnt     Score     Error   Units
TraceHeaderBenchmark.serialize                                   sample  463614     0.781 �   0.055   us/op
TraceHeaderBenchmark.serialize:serialize�p0.00                   sample             0.300             us/op
TraceHeaderBenchmark.serialize:serialize�p0.50                   sample             0.500             us/op
TraceHeaderBenchmark.serialize:serialize�p0.90                   sample             0.900             us/op
TraceHeaderBenchmark.serialize:serialize�p0.95                   sample             1.200             us/op
TraceHeaderBenchmark.serialize:serialize�p0.99                   sample             1.700             us/op
TraceHeaderBenchmark.serialize:serialize�p0.999                  sample            18.880             us/op
TraceHeaderBenchmark.serialize:serialize�p0.9999                 sample           119.964             us/op
TraceHeaderBenchmark.serialize:serialize�p1.00                   sample          2568.192             us/op
TraceHeaderBenchmark.serialize:�gc.alloc.rate                    sample      15  1196.732 � 300.700  MB/sec
TraceHeaderBenchmark.serialize:�gc.alloc.rate.norm               sample      15  1136.309 �  12.570    B/op
TraceHeaderBenchmark.serialize:�gc.churn.G1_Eden_Space           sample      15  1205.300 � 315.077  MB/sec
TraceHeaderBenchmark.serialize:�gc.churn.G1_Eden_Space.norm      sample      15  1143.188 �  73.104    B/op
TraceHeaderBenchmark.serialize:�gc.churn.G1_Survivor_Space       sample      15     0.044 �   0.015  MB/sec
TraceHeaderBenchmark.serialize:�gc.churn.G1_Survivor_Space.norm  sample      15     0.045 �   0.023    B/op
TraceHeaderBenchmark.serialize:�gc.count                         sample      15   126.000            counts
TraceHeaderBenchmark.serialize:�gc.time                          sample      15   217.000                ms
```

Before
```
Benchmark                                                          Mode     Cnt     Score     Error   Units
TraceHeaderBenchmark.serialize                                   sample  422292     0.932 �   0.063   us/op
TraceHeaderBenchmark.serialize:serialize�p0.00                   sample             0.400             us/op
TraceHeaderBenchmark.serialize:serialize�p0.50                   sample             0.700             us/op
TraceHeaderBenchmark.serialize:serialize�p0.90                   sample             1.000             us/op
TraceHeaderBenchmark.serialize:serialize�p0.95                   sample             1.400             us/op
TraceHeaderBenchmark.serialize:serialize�p0.99                   sample             1.900             us/op
TraceHeaderBenchmark.serialize:serialize�p0.999                  sample            20.992             us/op
TraceHeaderBenchmark.serialize:serialize�p0.9999                 sample           133.738             us/op
TraceHeaderBenchmark.serialize:serialize�p1.00                   sample          3518.464             us/op
TraceHeaderBenchmark.serialize:�gc.alloc.rate                    sample      15  1445.319 � 337.310  MB/sec
TraceHeaderBenchmark.serialize:�gc.alloc.rate.norm               sample      15  1664.383 �  12.583    B/op
TraceHeaderBenchmark.serialize:�gc.churn.G1_Eden_Space           sample      15  1459.855 � 339.541  MB/sec
TraceHeaderBenchmark.serialize:�gc.churn.G1_Eden_Space.norm      sample      15  1683.273 �  85.850    B/op
TraceHeaderBenchmark.serialize:�gc.churn.G1_Survivor_Space       sample      15     0.044 �   0.025  MB/sec
TraceHeaderBenchmark.serialize:�gc.churn.G1_Survivor_Space.norm  sample      15     0.051 �   0.025    B/op
TraceHeaderBenchmark.serialize:�gc.count                         sample      15   133.000            counts
TraceHeaderBenchmark.serialize:�gc.time                          sample      15   226.000                ms
```